LUT in progress

+++ src/spiky/lut/tests/test_lut_transformer_product.py
@@ -442,7 +442,7 @@ def main():
         devices.append('cuda:5')
 
     for device in devices:
-        for summation_dtype in [torch.int32]:  # , torch.int32
+        for summation_dtype in [torch.float32]:  # , torch.int32
             print(f"\nTesting on {device}, summation_dtype {summation_dtype}...")
             success = test_lut_transformer_product(device, summation_dtype, seed=123)
